Update: Add Structure for the mdns impl
