+++ .vscode/c_cpp_properties.json
@@ -1,29 +1,30 @@
 {
-  "configurations": [
-    {
-      "windowsSdkVersion": "10.0.19041.0",
-      "compilerPath": "C:/TDM-GCC-64/bin/g++.exe",
-      "name": "Win32",
-      "includePath": [
-        "C:/Program Files/OpenSSL-Win64/include/**",
-        "${workspaceFolder}/",
-        "${workspaceFolder}/build/_deps/**",
-        "C:/Qt/6.5.1/mingw_64/include/**",
-        "C:/Qt/6.5.1/mingw_64/include/QtGui",
-        "${workspaceFolder}/build/_deps/singleapplication-src",
-        "C:/Qt/6.5.1/mingw_64/include/QtCore",
-        "${workspaceFolder}/build"
-      ],
-      "defines": [
-        "_DEBUG",
-        "UNICODE",
-        "_UNICODE"
-      ],
-      "cStandard": "c17",
-      "cppStandard": "c++20",
-      "intelliSenseMode": "windows-gcc-x64",
-      "configurationProvider": "ms-vscode.cpptools"
-    }
-  ],
-  "version": 4
+    "configurations": [
+        {
+            "windowsSdkVersion": "10.0.19041.0",
+            "compilerPath": "C:/TDM-GCC-64/bin/g++.exe",
+            "name": "Win32",
+            "includePath": [
+                "C:/Program Files/OpenSSL-Win64/include/**",
+                "${workspaceFolder}/",
+                "${workspaceFolder}/build/_deps/**",
+                "C:/Qt/6.5.1/mingw_64/include/**",
+                "C:/Qt/6.5.1/mingw_64/include/QtGui",
+                "${workspaceFolder}/build/_deps/singleapplication-src",
+                "C:/Qt/6.5.1/mingw_64/include/QtCore",
+                "${workspaceFolder}/build",
+                "C:/Program Files/Bonjour SDK/Include"
+            ],
+            "defines": [
+                "_DEBUG",
+                "UNICODE",
+                "_UNICODE"
+            ],
+            "cStandard": "c17",
+            "cppStandard": "c++20",
+            "intelliSenseMode": "windows-gcc-x64",
+            "configurationProvider": "ms-vscode.cpptools"
+        }
+    ],
+    "version": 4
 }
