+++ docs/design/Protocol.md
@@ -17,7 +17,7 @@ This architecture ensures efficient and organized communication between devices,
 
 ## Discovering Devices
 
-To enable clipboard synchronization between devices on the same network, the Clipbird app utilizes a client-server approach. Initially, there is a server and one or more clients. When a client launches the app, it sends a Broadcast packet across the local network, seeking a potential server. If there is a server present within the network, it will respond with its details, allowing the client to establish a connection with the identified server. Once the connection is established, the clipboard syncing process can begin, facilitating the seamless sharing of clipboard data between the server and all connected clients on the same local network. This discovery mechanism ensures that devices can easily find and connect with a compatible server, optimizing the clipboard synchronization experience within the local network environment.
+Clipbird utilizes a discovery mechanism to identify compatible devices within the local network by utilizing the mdns protocol. This mechanism allows the client to discover the server without requiring the user to manually enter the IP address and port number.
 
 ## Protocol
 
@@ -33,11 +33,9 @@ Clipbird utilizes a variety of packet types for different purposes. These packet
 
 ### What are the packets Required for Clipbird
 
-During the synchronization process, two types of packets are used. The first type is the **Server Discovery Request** packet, which the client sends to initiate communication. This packet prompts the server to respond with a message in the form of a **Server Discovery Response** packet. This exchange helps the client locate the server within the local network. To simplify the process, we can note that both packets have the same structure. Therefore, we can combine them into a single type of packet called the **DiscoveryPacket**.
-
 Once the server has been identified, clipboard data is transmitted between the client and the server using a single type of packet known as the **SyncingPacket**. This packet is responsible for transferring clipboard data from the client to the server and vice versa, ensuring seamless sharing of clipboard content between the two devices.
 
-Finally we have **InvalidRequest** which is used to indicate that the packet sent by client is invalid so it provides a way to indicate that the packet is invalid. This packet should only sent from server to client not from client to server.
+Finally we have **InvalidRequest** which is used to indicate that the packet sent by client is invalid so it provides a way to indicate that the packet status. This packet should only sent from server to client not from client to server.
 
 ### InvalidRequest
 
@@ -68,31 +66,6 @@ The **InvalidRequest** is used to indicate that the packet is invalid. This pack
 |------------|---------------|
 | 0x01       | Coding Error  |
 | 0x02       | TLS Error     |
-
-### DiscoveryPacket
-
-The **DiscoveryPacket** packet is sent by the client to discover a compatible server within the local network. This packet contains the following fields:
-
-#### Header
-
-- **Packet Type**: This field specifies the type of packet, which is set to 0x01 for the Server Discovery Request and 0x02 for the Server Discovery Response.
-- **Packet Length**: This field specifies the length of the packet, for server discovery request packet it is length of ip address of host and port number.
-
-#### Body
-
-- **IP Type**: This field specifies the type of IP address, which can be IPv4 (0x04) or IPv6 (0x06).
-- **host IP**: This field contains the IP address of the host device.
-- **host Port**: This field contains the port number of the host device.
-
-#### Structure
-
-| Field           | Bytes  | value     |
-|-----------------|--------|-----------|
-| Packet Type     | 1      | 0x01/0x02 |
-| Packet Length   | 4      |           |
-| IP Type         | 1      | 0x04/0x06 |
-| host IP         | varies |           |
-| host Port       | 2      |           |
 
 ### SyncingPacket
