+++ tests/network/packets/SyncingPacket.hpp
@@ -11,6 +11,7 @@
 #include <QByteArray>
 
 // Local header files
+#include "network/packets/syncingitem/syncingitem.hpp"
 #include "network/packets/syncingpacket/syncingpacket.hpp"
 #include "types/enums/enums.hpp"
 #include "utility/functions/nbytes/nbytes.hpp"
