+++ tests/test.cpp
@@ -7,7 +7,6 @@
 #include <gtest/gtest.h>
 
 // Local header files
-#include "tests/network/packets/DiscoveryPacket.hpp"
 #include "tests/network/packets/InvalidRequest.hpp"
 #include "tests/network/packets/SyncingPacket.hpp"
