+++ utility/functions/packet/packet.cpp
@@ -6,42 +6,6 @@
 #include "packet.hpp"
 
 namespace srilakshmikanthanp::clipbirdesk::utility::functions {
-/**
- * @brief Create the DiscoveryPacket
- *
- * @param packetType
- * @param ipType
- * @param ipaddr
- * @param port
- *
- * @return DiscoveryPacket
- */
-network::packets::DiscoveryPacket createPacket(internals::DiscoveryPacketParams params) {
-  // create the packet
-  network::packets::DiscoveryPacket packet;
-
-  // set the packet type
-  packet.setPacketType(params.packetType);
-
-  // set the ip type
-  packet.setIpType(params.ipType);
-
-  // set the ip address
-  if (params.ipType == types::enums::IPType::IPv4) {
-    packet.setHostIp(toIPV4QByteArray(params.ipaddr));
-  } else {
-    packet.setHostIp(toIPV6QByteArray(params.ipaddr));
-  }
-
-  // set the port
-  packet.setHostPort(params.port);
-
-  // set the packet length
-  packet.setPacketLength(packet.size());
-
-  // return the packet
-  return packet;
-}
 
 /**
  * @brief Create the ErrorMessage
